Verify gpcheckcat columns in Python cluster unit tests.
